src: move getActiveResources/Handles to node_process.cc
